+++ Documentation/Changelog.md
@@ -11,6 +11,7 @@ Both primitives provide drop-in replacements for `System.Threading.Tasks.Paralle
 - Added ability to create LINQ expressions from functions via `Gapotchenko.FX.Fn` primitive
 - Implemented polyfill for `Enumerable.ToHashSet<T>` operation
 - Implemented polyfills for `BitConverter.SingleToInt32Bits` and `Int32BitsToSingle` operations
+- Implemented polyfills to the future for `WaitForExit(CancellationToken)` and `WaitForExit(int, CancellationToken)` methods of `System.Diagnostics.Process` type
 - `Empty.Task` is not suggested by the code editor when it is natively provided by the host platform
 - Fixed issue with ambiguous match of `Append` and `Prepend` polyfills for `IEnumerable<T>` type of some target frameworks
 - Fixed issue with binding redirects handling in `Gapotchenko.FX.Reflection.Loader` module that could lead to `StackOverflowException` under specific conditions